Periodicity works now in 4.3 and 5 due to _0 part in datefield
Different TAL define attribute processing caused problem, as Plone 4.3
adds _0 to formfield.
Form entry was missing and now restored back in
booking_periodicity_ajax_form.  And show_hm: False: is deleted asl we
want the time to be visible (again). It van have effect on the last date
of booking a periodic. You can influence the outcome. without this time
field it was not possible.